dfsm: Change the brackets used for array construction
Change them from parentheses to square brackets, in line with the Python
syntax for D-Bus arrays. This introduces a reduce–reduce conflict between
array construction and indexing, so syntax support for array indexing has
been removed. It’s an irregularly used operation, so can be provided by a
built-in function instead.