dfsm: Split AST node checking methods
Split check() into sanity_check(), pre_check_and_register() and check(). The
first is for sanity checks of the state of a node at any time during the
interpretation process, and should always pass (or the program will abort).
The second is for checks which can be performed without needing to read
information from the environment; if it passes, it adds any relevant
information to the environment. The final method is for any remaining checks
which require information from the environment.