dfsm: Fix type calculation for complex data structures
Previously, type calculation would fail if (for example) any of the elements
of an array differed in type. Now, in this case, we choose the type of the
array to be the least-general supertype of the types of its elements so that
this can't happen. In practice, this produces arrays of variants.

The same change has been made to dictionaries.