dfsm: Add syntax support for variant data structures
Since a{sv} is so popular, we can't avoid adding explicit support for variant
data structures in the language. We use the same syntax as GVariant's print
format, i.e. variants are surrounded by angle brackets: “<"foo">”.

This necessitates changing the symbols for less-than and greater-than to
‘<~’ and ‘~>’. A bit ugly, but I couldn't think of anything better which
wouldn't be ambiguous.