dfsm: Parse D-Bus object paths and type signatures as strings
In line with the GVariant text format, we now parse D-Bus object paths and
type signatures as strings (i.e. they have to be quoted). This should reduce
the awkwardness of the compiler considerably, since type signatures messed
around with a lot of things.

This means that currently it isn't possible to create an object path or type
signature, since they'll always be treated as strings. Type annotations are
coming soon which will fix that.