dfsm: Add type annotations to the language
In the same style as GVariant's text format type annotations (and using the
same syntax), we solve the problem of determining the type of (e.g.) an
empty array by requiring it to have a type annotation.

Coupled with the support for variants, this means we can change the default
least-general supertype code to use an any type as the least-general
supertype, rather than a variant.