dfsm: Reduce the lengths of whitespace prefixes/suffixes generated for strings
The capability of whitespace to trigger bugs is most likely in its mere
presence, rather than much to do with the amount of whitespace present.
Consequently, it’s just an annoyance to have up to 50 characters of whitespace
before and after fuzzed strings. Reduce this to 5. This strikes a balance
between having enough whitespace to contain a reasonable selection of
different whitespace characters, and not being too annoying.

Perhaps a better approach would be to have a non-uniform probability
distribution giving the whitespace length, but I can’t be bothered to write
one at the moment.