dfsm: Remove syntactical regexp support
The original intention with regexps was to allow fuzzy strings to be
generated which were accepted by the regexps, allowing a (relatively)
concise way of fuzzing structured data.

However, this need is fairly well covered by default values and fuzzing of
structural strings (by treating them as blocks of text).

Generating strings to match a given regexp is hard, though there are
libraries to achieve it. At this point, it's not worth putting the time into
getting that working. I've got other deadlines to meet.