dfsm: Trim error handling
It turns out that since we're especially strict about typing data structures,
we can trim a lot of GErrors from data structures and expressions, to the
point where we can guarantee zero runtime errors from anything smaller than
a statement.

At the moment, the only reason to keep GErrors in statement execution methods
is because they carry D-Bus error information up the stack. This is ugly and
needs to be killed.