dfsm: Re-work how D-Bus signal emissions & method replies are handled
We now pass a DfsmOutputSequence object down to the DfsmMachine, and all
the effects of its statements are appended to a list in the
DfsmOutputSequence. The caller can then evaluate the output sequence.

This allows the caller to have complete control over what actually happens
when emitting a D-Bus signal, or responding to a D-Bus method call, and
means it’s all handled nicely inside DfsmObject.

This also solves the problem whereby code like “reply; emit” would cause the
signal emission to happen before the D-Bus method reply, which was breaking
the EDS address book machine.