machines: Add some more structure to eds-address-book.machine’s vCards
Built the fuzzy vCard strings from a fuzzed array of valid vCard properties.

This assumes that EDS’ vCard parser is very reliable when it comes to
handling invalidly-structured vCards. Instead, we now concentrate on slipping
invalid property values (or invalid combinations of properties) past the
parser and deeper into EDS.