bendy-bus: Wait for test program to die before stopping the simulation
When we want to stop the simulation, we need to send SIGTERM to the test
program, _and then wait for it to die_ before we continue shutting down the
simulation. This is because test programs may handle SIGTERM and spend a
while shutting themselves down. Evolution does, for example.

If programs do that but we continue to shut down without waiting for them,
when running bendy-bus-lcov, we'll end up trying to process the test
program's .gcda files before the program has written them yet.