bendy-bus: Add a SIGKILL on a timeout when stopping the simulation
Sometimes the test program won’t respond to SIGTERM. Since we now wait for
notification that the child process has died, the simulation will continue
running forever in this case. This is bad. Consequently, we now send a
SIGKILL to the child process after a 15 second timeout if it doesn’t respond
to SIGTERM.

This does mean that the process won’t (for example) write out code coverage
files (since a process has to exit cleanly to do that), but in this situation
we can never hope to get those anyway. It’s better to kill the process and
just start on another test run than continue on the current one indefinitely.