dfsm: Use type annotations to give integer types
This removes the ugly postfix syntax for specifying the types of integers,
and uses type annotations instead. This is much more like the GVariant text
format, in that all unannotated integers are presumed to be int32s, and
annotations are used otherwise to specify numeric types.

The only differences now are that libdfsm doesn't support:
 • Full-length type annotations (e.g. ‘uint32 5’ vs. ‘@u 5’)
 • Type inference from surrounding types (e.g. ‘@au [5]’)